chore(linting): use airbnb setup
