+++ .eslintrc.js
@@ -1,0 +1,30 @@
+module.exports = {
+  env: {
+    browser: true,
+    es6: true,
+  },
+  extends: [
+    'airbnb-base',
+  ],
+  globals: {
+    Atomics: 'readonly',
+    SharedArrayBuffer: 'readonly',
+  },
+  parser: '@typescript-eslint/parser',
+  parserOptions: {
+    ecmaVersion: 2018,
+    sourceType: 'module',
+  },
+  plugins: [
+    '@typescript-eslint', 'jest'
+  ],
+  settings: {
+    "import/resolver": {
+      node: {
+        extensions: [".js", ".ts"]
+      }
+    }
+  },
+  rules: {
+  },
+};
