+++ src/errors/Custom.ts
@@ -1,0 +1,9 @@
+export class CustomError extends Error {
+  constructor(message?: string) {
+    super(message);
+    this.name = this.constructor.name;
+    Error.captureStackTrace(this, this.constructor);
+  }
+}
+
+export default { CustomError };
