+++ src/errors/index.ts
@@ -1,0 +1,11 @@
+import { BlacklistedIPError } from './BlacklistedIP';
+import { DNSResolveError } from './DNSResolve';
+import { MalformedURLError } from './MalformedURL';
+import { NetworkError } from './Network';
+
+export {
+  BlacklistedIPError, DNSResolveError, MalformedURLError, NetworkError,
+};
+export default {
+  BlacklistedIPError, DNSResolveError, MalformedURLError, NetworkError,
+};
